Modified a few components.
Added a pre-emptive main loop.
Added a point helper function to the rays class so as to get a point of
the ray with a float value passed to it. (The origin*direction*t)
Considering using std::pair from tuple to extract values.
Color needs to be computed using Phong, TO DO